fix_circle_dependencies
